# docs/swagger.yaml
definitions:
  model.AccessToken:
    properties:
      token:
        type: string
    type: object
  model.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  model.Event:
    properties:
      data:
        $ref: '#/definitions/model.EventData'
      timestamp:
        type: string
      type:
        type: string
    type: object
  model.EventData:
    properties:
      gameId:
        type: integer
      payload:
        type: string
      playerId:
        type: integer
    type: object
...
  model.GenericResponse:
    properties:
      data:
        type: string
      success:
        type: boolean
    type: object
  model.Player:
    properties:
      createdAt:
        type: string
      draws:
        type: integer
      elo:
        type: integer
      id:
        type: integer
      isPlaying:
        type: boolean
      lastPlayedAt:
        type: string
      losses:
        type: integer
      rate:
        type: number
      username:
        type: string
      wins:
        type: integer
    type: object
...
  model.PlayerRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: lukamatosevic5@gmail.com
    name: Luka Matošević
    url: https://lukamatosevic.com
  license:
    name: MIT 2023
    url: https://www.mit.edu/~amini/LICENSE.md
paths:
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login registered player
      parameters:
      - description: Login player
        in: body
        name: player
        required: true
        schema:
          $ref: '#/definitions/model.PlayerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.AccessToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Login registered player
      tags:
      - auth
  /v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout authorized player
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Logout authorized player
      tags:
      - auth
  /v1/auth/player:
    get:
      consumes:
      - application/json
      description: Get authorized player
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.Player'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get authorized player
      tags:
      - auth
  /v1/events/subscribe:
    get:
      consumes:
      - application/json
      description: Subscribe to server sent events
      parameters:
      - description: Access token
        in: query
        name: token
        required: true
        type: string
      - description: Event type
        enum:
        - GameAnyEvent
        - GameMoveEvent
        - GameJoinEvent
        - GameQuitEvent
        - GameStartEvent
        - GameEndEvent
        - GameWhitePlayerMoveEvent
        - GameWhitePlayerMoveEvent
        in: query
        name: event
        required: true
        type: string
      - description: Game ID
        in: query
        name: gameId
        type: integer
      produces:
      - text/event-stream
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.Event'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Subscribe to server sent events
      tags:
      - events
...
securityDefinitions:
  ApiKeyAuth:
    description: The access token obtained from /login endpoint, required for accessing
      protected routes
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"